Eric Sunshine replaced fabsf() with fabs() in terrbig plugin for improved portability since not all platforms have fabsf().  (An alternate solution would be to check for fabsf via configure.  We will probably want to do this in the future.)
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/migrated/branches/R0_96@12666 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/terrbig/nterrain.cpp
@@ -627,7 +627,7 @@ void csBigTerrainObject::ComputeLod (nBlock *heightmap, int i, int j, int di, in
   nBlock *b = &heightmap[i + j * width];
   nBlock *l = &heightmap[i - di + (j - dj) * width];
   nBlock *r = &heightmap[i + di + (j + dj) * width];
-  b->error = fabsf (b->pos.y - (l->pos.y + r->pos.y) / 2.0);
+  b->error = (float)fabs(b->pos.y - (l->pos.y + r->pos.y) / 2.0);
   if (n) {
     dj = (di + dj) / 2;
     di -= dj;
